+++ src/shapes/rect/irect.ts
@@ -24,11 +24,11 @@
  *  THE SOFTWARE.
  */
 
-declare module powerbi.extensibility.utils.svg {
+// declare module powerbi.extensibility.utils.svg {
     export interface IRect {
         left: number;
         top: number;
         width: number;
         height: number;
     }
-}
+// }