+++ test/manipulationTest.ts
@@ -23,62 +23,70 @@
  *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  *  THE SOFTWARE.
  */
+import { parseTranslateTransform, getTransformScaleRatios } from "../src/manipulation";
+import { select } from "d3-selection";
 
-module powerbi.extensibility.utils.svg.test {
-    // powerbi.extensibility.utils.svg
-    import parseTranslateTransform = powerbi.extensibility.utils.svg.parseTranslateTransform;
-    import getTransformScaleRatios = powerbi.extensibility.utils.svg.getTransformScaleRatios;
 
-    // powerbi.extensibility.utils.test
-    import testDom = powerbi.extensibility.utils.test.helpers.testDom;
+import * as TestUtils from "powerbi-visuals-utils-testutils";
 
-    describe("SvgUtil tests", () => {
-        it("validate the pie chart transform parsing logic for Chrome", () => {
-            let transform = "translate(110.21,46.5)";
-            let parsedTransform = parseTranslateTransform(transform);
 
-            expect(parsedTransform.x).toBe("110.21");
-            expect(parsedTransform.y).toBe("46.5");
-        });
+import { testDom } from "powerbi-visuals-utils-testutils";
 
-        it("validate the pie chart transform parsing logic for IE", () => {
-            let transform = "translate(110.6 34.56)";
-            let parsedTransform = parseTranslateTransform(transform);
 
-            expect(parsedTransform.x).toBe("110.6");
-            expect(parsedTransform.y).toBe("34.56");
-        });
+// module powerbi.extensibility.utils.svg.test {
+// powerbi.extensibility.utils.svg
+// powerbi.extensibility.utils.test
+describe("SvgUtil tests", () => {
+    it("validate the pie chart transform parsing logic for Chrome", () => {
+        let transform = "translate(110.21,46.5)";
+        let parsedTransform = parseTranslateTransform(transform);
 
-        it("validate transform parsing logic with no y value", () => {
-            let transform = "translate(110.6)";
-            let parsedTransform = parseTranslateTransform(transform);
-
-            expect(parsedTransform.x).toBe("110.6");
-            expect(parsedTransform.y).toBe("0");
-        });
-
-        it("get transform scale ratios under parent scope", (done) => {
-            let jqDiv = testDom("500", "500");
-            jqDiv.css("transform", "scale(0.75,0.5)");
-            let svg = d3.select(jqDiv[0]).append("svg").attr({
-                width: 350,
-                height: 200,
-            }).style("position", "absolute");
-            let g = svg.append("g"); // the axisGraphicsContext
-            g.append("rect").attr({
-                x: 0, y: 200, width: 350, height: 80,
-            }).style("fill", "red"); // this rect is simulating the x-axis which fills the <svg> parent
-            g.append("rect").attr({
-                x: 0, y: 200, width: 80, height: 200,
-            }).style("fill", "red"); // this rect is simulating the y-axis...
-
-            setTimeout(() => {
-                let ratios = getTransformScaleRatios(<SVGSVGElement>svg.node());
-                expect(ratios.x).toBe(0.75);
-                expect(ratios.y).toBe(0.5);
-
-                done();
-            }, 10);
-        });
+        expect(parsedTransform.x).toBe("110.21");
+        expect(parsedTransform.y).toBe("46.5");
     });
-}
+
+    it("validate the pie chart transform parsing logic for IE", () => {
+        let transform = "translate(110.6 34.56)";
+        let parsedTransform = parseTranslateTransform(transform);
+
+        expect(parsedTransform.x).toBe("110.6");
+        expect(parsedTransform.y).toBe("34.56");
+    });
+
+    it("validate transform parsing logic with no y value", () => {
+        let transform = "translate(110.6)";
+        let parsedTransform = parseTranslateTransform(transform);
+
+        expect(parsedTransform.x).toBe("110.6");
+        expect(parsedTransform.y).toBe("0");
+    });
+
+    it("get transform scale ratios under parent scope", (done) => {
+        let jqDiv = testDom("500", "500");
+        jqDiv.css("transform", "scale(0.75,0.5)");
+        let svg = select(jqDiv[0]).append("svg")
+            .attr("width", 350)
+            .attr("height", 200).style("position", "absolute");
+        let g = svg.append("g"); // the axisGraphicsContext
+        g.append("rect")
+            .attr("x", 0)
+            .attr("y", 200)
+            .attr("width", 350)
+            .attr("height", 80)
+            .style("fill", "red"); // this rect is simulating the x-axis which fills the <svg> parent
+        g.append("rect")
+            .attr("x", 0)
+            .attr("y", 200)
+            .attr("width", 80)
+            .attr("height", 200).style("fill", "red"); // this rect is simulating the y-axis...
+
+        setTimeout(() => {
+            let ratios = getTransformScaleRatios(<SVGSVGElement>svg.node());
+            expect(ratios.x).toBe(0.75);
+            expect(ratios.y).toBe(0.5);
+
+            done();
+        }, 10);
+    });
+});
+// }