+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0
+* Moved to webpack 3
+* Moved to commonjs style
+
+
 ## 1.1.0
 * Removed `typings`
 * Unified dependencies versions